refactor: Separate binary-operator-related classes
