+++ Includes/Operators/BinaryOperator.hpp
@@ -26,60 +26,4 @@ public:
 	virtual std::string GetInfix(bool = false) const;
 };
 
-class Multiply : public BinaryOperator
-{
-protected:
-	Multiply(const Multiply&);
-
-public:
-	Multiply(ExpressionTree* = 0, ExpressionTree* = 0);
-
-	ExpressionTree* Clone() const;
-	std::string ToString() const;
-	double Evaluate() const;
-};
-
-class Divide : public BinaryOperator
-{
-protected:
-	Divide(const Divide&);
-
-public:
-	Divide(ExpressionTree* = 0, ExpressionTree* = 0);
-
-	ExpressionTree* Clone() const;
-	std::string ToString() const;
-	double Evaluate() const;
-};
-
-class Power : public BinaryOperator
-{
-protected:
-	Power(const Power&);
-
-public:
-	Power(ExpressionTree* = 0, ExpressionTree* = 0);
-
-	ExpressionTree* Clone() const;
-	std::string ToString() const;
-	double Evaluate() const;
-};
-
-class Assign : public BinaryOperator
-{
-private:
-	VariableTable& m_varTable;
-
-protected:
-	Assign(const Assign&);
-
-public:
-	Assign(ExpressionTree*, ExpressionTree*, VariableTable&);
-
-	ExpressionTree* Clone() const;
-	std::string ToString() const;
-	double Evaluate() const;
-	virtual std::string GetInfix(bool = false) const;
-};
-
 #endif
