+++ Sources/Operators/BinaryOperator.cpp
@@ -48,126 +48,3 @@ void BinaryOperator::Print(std::ostream& os, int depth) const
 
 	m_leftTree->Print(os, depth + 3);
 }
-
-Multiply::Multiply(ExpressionTree* left, ExpressionTree* right)
-	: BinaryOperator(left, right)
-{ }
-
-Multiply::Multiply(const Multiply& multiplyNode)
-	: BinaryOperator(multiplyNode)
-{ }
-
-ExpressionTree* Multiply::Clone() const
-{
-	return new Multiply(*this);
-}
-
-std::string Multiply::ToString() const
-{
-	return "*";
-}
-
-double Multiply::Evaluate() const
-{
-	return m_leftTree->Evaluate() * m_rightTree->Evaluate();
-}
-
-Divide::Divide(ExpressionTree* left, ExpressionTree* right)
-	: BinaryOperator(left, right)
-{ }
-
-Divide::Divide(const Divide& divideNode)
-	: BinaryOperator(divideNode)
-{ }
-
-ExpressionTree* Divide::Clone() const
-{
-	return new Divide(*this);
-}
-
-std::string Divide::ToString() const
-{
-	return "/";
-}
-
-double Divide::Evaluate() const
-{
-	double tempEval = m_rightTree->Evaluate();
-
-	if (tempEval == 0)
-	{
-		throw ExpressionTreeError("Divided by 0");
-	}
-
-	return m_leftTree->Evaluate() / tempEval;
-}
-
-Power::Power(ExpressionTree* left, ExpressionTree* right)
-	: BinaryOperator(left, right)
-{ }
-
-Power::Power(const Power& powerNode)
-	: BinaryOperator(powerNode)
-{ }
-
-ExpressionTree* Power::Clone() const
-{
-	return new Power(*this);
-}
-
-std::string Power::ToString() const
-{
-	return "^";
-}
-
-double Power::Evaluate() const
-{
-	return std::pow(m_leftTree->Evaluate(), m_rightTree->Evaluate());
-}
-
-Assign::Assign(ExpressionTree* left, ExpressionTree* right, VariableTable& varTable)
-	: BinaryOperator(left, right), m_varTable(varTable)
-{ }
-
-Assign::Assign(const Assign& assignNode)
-	: BinaryOperator(assignNode), m_varTable(assignNode.m_varTable)
-{ }
-
-ExpressionTree* Assign::Clone() const
-{
-	return new Assign(*this);
-}
-
-std::string Assign::ToString() const
-{
-	return "=";
-}
-
-std::string Assign::GetInfix(bool) const
-{
-	return m_leftTree->GetInfix(true) + " " + ToString() + " " + m_rightTree->GetInfix(true);
-}
-
-double Assign::Evaluate() const
-{
-	double rightTreeValue = m_rightTree->Evaluate();
-	Variable* varPointer = dynamic_cast<Variable*>(m_leftTree);
-
-	if (varPointer)
-	{
-		if (m_varTable.Find(varPointer->ToString()))
-		{
-			varPointer->SetValue(rightTreeValue);
-		}
-		else
-		{
-			m_varTable.Insert(varPointer->ToString(), rightTreeValue);
-		}
-	}
-	else
-	{
-		throw ExpressionTreeError("Invalid Assign");
-	}
-
-	return rightTreeValue;
-}
